requirements.txt
# IMAP and Email Processing
imapclient==3.0.1
email-validator==2.1.0
dkimpy==1.1.5
pyspf==2.0.14
dnspython==2.6.1

# Database
mysql-connector-python==8.3.0
sqlalchemy>=2.0.32
bcrypt>=4.0.0

# Web Framework
flask==3.0.2
flask-cors==4.0.0

# API Clients
anthropic>=0.40.0
openai>=1.12.0  # Used for both OpenAI and Perplexity APIs
requests==2.31.0

# OSINT and Analysis
python-whois==0.9.4
ipwhois==1.1.0
geoip2==4.7.0
beautifulsoup4==4.12.3
lxml>=5.3.0

# PDF Generation
reportlab==4.1.0
weasyprint==61.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
jinja2==3.1.3
pillow>=10.4.0

# Background Tasks
schedule==1.2.1

# Logging
colorlog==6.8.2
